price_crawler/requirements.txt
confluent-kafka==2.6.1
protego==0.3.1
requests==2.32.3
selenium==4.27.1
beautifulsoup4==4.12.3
jsonschema==4.23.0
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
pytest==8.3.4
google-generativeai==0.8.3
googlesearch-python==1.2.5
google==3.0.0